More Pages

+++ admin/delete-category.php
@@ -8,12 +8,16 @@ if (isset($_GET['id'])) {
     //for later///////////////////////
 
     //update category_id of posts that belong to thid category to id of uncategorized category
+    $update_query = "UPDATE posts SET category_id = 7 where category_id = $id";
+    $update_result = mysqli_query($connection, $update_query);
 
-
+    if (!mysqli_errno($connection)) {
     //delete category
     $query = " DELETE FROM categories WHERE id = $id limit 1";
     $result = mysqli_query($connection, $query);
     $_SESSION['delete-category-success'] = "Category deleted successfully";
+    }
+    
 }
 header('location: ' . ROOT_URL .'admin/manage-categories.php');
 die();
